fix: move validation to separate folder
