+++ src/store/validation/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Store validation - re-export validation functions
+ */
+export {
+  validateTransaction,
+  validateRecurringTransaction,
+} from './finance.validation';
